requirements.txt
pydantic~=1.10.2
setuptools~=65.5.1
cryptography~=38.0.3
jsonschema~=4.17.0
canonicaljson~=1.6.4
requests~=2.28.1
uvicorn~=0.19.0
fastapi~=0.86.0
starlette~=0.21.0
passlib~=1.7.4
SQLAlchemy~=1.4.43
python-multipart~=0.0.5
python-jose[cryptography]~=3.3.0
bcrypt~=4.0.1
python-snappy~=0.6.1
numpy~=1.23.4